move design be language and audio code into shared libraries
