+++ apps/design/backend/package.json
@@ -29,8 +29,6 @@
     "type-check": "tsc --build"
   },
   "dependencies": {
-    "@google-cloud/text-to-speech": "^5.0.1",
-    "@google-cloud/translate": "^8.0.2",
     "@votingworks/backend": "workspace:*",
     "@votingworks/basics": "workspace:*",
     "@votingworks/db": "workspace:*",
@@ -48,7 +46,6 @@
     "fs-extra": "11.1.1",
     "js-sha256": "^0.9.0",
     "jszip": "^3.9.1",
-    "node-html-parser": "^6.1.13",
     "react": "18.3.1",
     "uuid": "9.0.1",
     "zod": "3.23.5"
