+++ libs/backend/package.json
@@ -24,6 +24,8 @@
     "type-check": "tsc --build"
   },
   "dependencies": {
+    "@google-cloud/text-to-speech": "^5.0.1",
+    "@google-cloud/translate": "^8.0.2",
     "@types/uuid": "9.0.5",
     "@votingworks/auth": "workspace:*",
     "@votingworks/basics": "workspace:*",
@@ -43,6 +45,7 @@
     "jszip": "^3.9.1",
     "luxon": "^3.0.0",
     "micromatch": "^4.0.5",
+    "node-html-parser": "^6.1.13",
     "readline": "^1.3.0",
     "stream-chain": "^2.2.5",
     "stream-chopper": "^3.0.1",
