+++ libs/backend/src/language_and_audio/google_cloud_config.ts
@@ -1,1 +1,2 @@
+/** Google Cloud Project ID */
 export const GOOGLE_CLOUD_PROJECT_ID = 'astral-pursuit-395520';
