+++ libs/backend/src/language_and_audio/index.ts
@@ -1,7 +1,10 @@
+/* istanbul ignore file */
 export * from './app_strings';
+export * from './ballot_strings';
 export * from './audio';
 export * from './election_strings';
 export * from './speech_synthesizer';
+export * from './test_utils';
 export * from './translator';
 export * from './utils';
 export * from './vendored_translations';
