+++ libs/backend/src/language_and_audio/test_utils.ts
@@ -1,0 +1,69 @@
+/* eslint-disable max-classes-per-file */
+/* eslint-disable vx/gts-jsdoc */
+/** istanbul ignore file */
+
+import { MinimalGoogleCloudTextToSpeechClient } from './speech_synthesizer';
+import { MinimalGoogleCloudTranslationClient } from './translator';
+
+export function mockCloudTranslatedText(
+  englishText: string,
+  languageCode: string
+): string {
+  return `${englishText} (in ${languageCode})`;
+}
+
+export class MockGoogleCloudTranslationClient
+  implements MinimalGoogleCloudTranslationClient
+{
+  // eslint-disable-next-line vx/gts-no-public-class-fields
+  translateText = jest.fn(
+    (input: {
+      contents: string[];
+      targetLanguageCode: string;
+    }): Promise<
+      [
+        { translations: Array<{ translatedText: string }> },
+        undefined,
+        undefined,
+      ]
+    > =>
+      Promise.resolve([
+        {
+          translations: input.contents.map((text) => ({
+            translatedText: mockCloudTranslatedText(
+              text,
+              input.targetLanguageCode
+            ),
+          })),
+        },
+        undefined,
+        undefined,
+      ])
+  );
+}
+
+export function mockCloudSynthesizedSpeech(text: string): string {
+  return `${text} (audio)`;
+}
+
+export function isMockCloudSynthesizedSpeech(audioContent: string): boolean {
+  return audioContent.endsWith(' (audio)');
+}
+
+export class MockGoogleCloudTextToSpeechClient
+  implements MinimalGoogleCloudTextToSpeechClient
+{
+  // eslint-disable-next-line vx/gts-no-public-class-fields
+  synthesizeSpeech = jest.fn(
+    (input: {
+      input: { text: string };
+    }): Promise<
+      [{ audioContent: string | Uint8Array }, undefined, undefined]
+    > =>
+      Promise.resolve([
+        { audioContent: mockCloudSynthesizedSpeech(input.input.text) },
+        undefined,
+        undefined,
+      ])
+  );
+}
