+++ libs/backend/src/language_and_audio/utils.test.ts
@@ -1,4 +1,8 @@
-import { UiStringAudioIdsPackage, UiStringsPackage } from '@votingworks/types';
+import {
+  UiStringAudioIdsPackage,
+  UiStringsPackage,
+  LanguageCode,
+} from '@votingworks/types';
 import {
   cleanText,
   forEachUiString,
@@ -8,7 +12,6 @@ import {
   setUiStringAudioIds,
   splitInterpolatedText,
 } from './utils';
-import { LanguageCode } from '../language_code';
 
 test.each<{ input: string; expectedOutput: string }>([
   {
