+++ libs/backend/src/language_and_audio/vendored_translations.test.ts
@@ -1,3 +1,4 @@
+import { assertDefined } from '@votingworks/basics';
 import { parseVendoredTranslations } from './vendored_translations';
 
 function areSetsEqual<T>(set1: Set<T>, set2: Set<T>): boolean {
@@ -23,6 +24,6 @@ test('vendored_translations.json', () => {
     .map((keys) => new Set(keys));
   const firstKeySet = keySetsForEachLanguage[0];
   for (const keySet of keySetsForEachLanguage) {
-    expect(areSetsEqual(firstKeySet, keySet)).toEqual(true);
+    expect(areSetsEqual(assertDefined(firstKeySet), keySet)).toEqual(true);
   }
 });
