+++ libs/backend/src/language_and_audio/vendored_translations.ts
@@ -1,9 +1,15 @@
 import { z } from 'zod';
-import { safeParse } from '@votingworks/types';
+import {
+  safeParse,
+  NonEnglishLanguageCode,
+  LanguageCode,
+} from '@votingworks/types';
 
 import vendoredTranslations from './vendored_translations.json';
-import { NonEnglishLanguageCode, LanguageCode } from '../language_code';
 
+/**
+ * A mapping of non-English language codes to translations of English text.
+ */
 export type VendoredTranslations = Record<
   NonEnglishLanguageCode,
   { [englishText: string]: string }
@@ -15,6 +21,9 @@ const VendoredTranslationsSchema: z.ZodSchema<VendoredTranslations> = z.object({
   [LanguageCode.SPANISH]: z.record(z.string()),
 });
 
+/**
+ * Parse the vendored translations from the JSON file.
+ */
 export function parseVendoredTranslations(): VendoredTranslations {
   return safeParse(
     VendoredTranslationsSchema,
