+++ libs/hmpb/src/hmpb_strings.tsx
@@ -1,41 +1,7 @@
 import React from 'react';
 import { mapObject } from '@votingworks/basics';
 import { UiString } from '@votingworks/ui';
-
-export const hmpbStringsCatalog = {
-  hmpbOfficialBallot: 'Official Ballot',
-  hmpbSampleBallot: 'Sample Ballot',
-  hmpbTestBallot: 'Test Ballot',
-  hmpbOfficialAbsenteeBallot: 'Official Absentee Ballot',
-  hmpbSampleAbsenteeBallot: 'Sample Absentee Ballot',
-  hmpbTestAbsenteeBallot: 'Test Absentee Ballot',
-  hmpbOfficialProvisionalBallot: 'Official Provisional Ballot',
-  hmpbSampleProvisionalBallot: 'Sample Provisional Ballot',
-  hmpbTestProvisionalBallot: 'Test Provisional Ballot',
-  hmpbInstructions: 'Instructions',
-  hmpbInstructionsToVoteTitle: 'To Vote:',
-  hmpbInstructionsToVoteText:
-    'To vote, completely fill in the oval next to your choice.',
-  hmpbInstructionsWriteInTitle: 'To Vote for a Write-in:',
-  hmpbInstructionsWriteInText:
-    'To vote for a person whose name is not on the ballot, write the person’s name on the "Write-in" line and completely fill in the oval next to the line.',
-  hmpbVoteFor1: 'Vote for 1',
-  hmpbVoteFor2: 'Vote for up to 2',
-  hmpbVoteFor3: 'Vote for up to 3',
-  hmpbVoteFor4: 'Vote for up to 4',
-  hmpbVoteFor5: 'Vote for up to 5',
-  hmpbVoteFor6: 'Vote for up to 6',
-  hmpbVoteFor7: 'Vote for up to 7',
-  hmpbVoteFor8: 'Vote for up to 8',
-  hmpbVoteFor9: 'Vote for up to 9',
-  hmpbVoteFor10: 'Vote for up to 10',
-  hmpbWriteIn: 'Write-in',
-  hmpbPageIntentionallyBlank: 'This page intentionally left blank',
-  hmpbContinueVotingOnBack: 'Turn ballot over and continue voting',
-  hmpbContinueVotingOnNextSheet: 'Continue voting on next ballot sheet',
-  hmpbVotingComplete: 'You have completed voting.',
-  hmpbPage: 'Page',
-} satisfies Record<string, string>;
+import { hmpbStringsCatalog } from '@votingworks/utils';
 
 export const hmpbStrings = mapObject(hmpbStringsCatalog, (string, key) => (
   <UiString uiStringKey={key}>{string}</UiString>
