+++ libs/types/src/ballot_language_config.ts
@@ -1,0 +1,17 @@
+import { LanguageCode } from './language_code';
+
+export interface BallotLanguageConfig {
+  languages: LanguageCode[];
+}
+
+export type BallotLanguageConfigs = BallotLanguageConfig[];
+
+export function getAllBallotLanguages(
+  ballotLanguageConfigs: BallotLanguageConfigs
+): LanguageCode[] {
+  const uniqueLanguages = new Set(
+    ballotLanguageConfigs.flatMap((b) => b.languages)
+  );
+
+  return [...uniqueLanguages];
+}
