+++ libs/types/src/index.ts
@@ -29,6 +29,8 @@ export * from './image';
 export * from './interpretation';
 export * from './ipp_printing';
 export * from './languages';
+export * from './language_code';
+export * from './ballot_language_config';
 export * from './numeric';
 export * from './polls';
 export * from './precinct_scanner';
