+++ libs/utils/src/hmpb/index.ts
@@ -1,1 +1,2 @@
 export * from './all_contest_options';
+export * from './hmpb_strings';
